♻️ Extract opts package with OptionType and OptionTypes
