+++ opts/option_types.go
@@ -1,3 +1,3 @@
-package main
+package opts
 
 type OptionTypes[T any] []*OptionType[T]
